+++ main.go
@@ -10,11 +10,22 @@ import (
 //go:embed abbreviation.json
 var abbreviations string
 
+var userDefinitionBytes []byte
+var userDefinitions *string
+
 func main() {
 	args := os.Args
 
 	if len(args) <= 1 {
-		helpers.Error("you did not enter a command to execute")
+		helpers.Error("you did not enter a command to execute, run abbreviation help for a list of commands")
+	}
+
+	if helpers.FileExists(helpers.GetConfigPath()) {
+		userDefinitionBytes, _ = os.ReadFile(helpers.GetConfigPath())
+		str := string(userDefinitionBytes)
+		userDefinitions = &str
+	} else {
+		userDefinitions = nil
 	}
 
 	cmd := args[1]
@@ -26,8 +37,14 @@ func main() {
 		case "help":
 			commands.Help(args)
 		case "search":
-			commands.Search(args, abbreviations)
+			commands.Search(args, abbreviations, userDefinitions)
 		case "list":
 			commands.List(abbreviations)
+		case "add":
+			commands.Add(args)
+		case "remove":
+			commands.Remove(args)
+		case "list-user":
+			commands.ListUser(userDefinitions)
 	}
 }