+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# v0.6.0
+
+Custom abbreviations are now supported! <br>
+The new commands save the custom abbreviations in these directories:
+
+- Windows: C:/Users/(your user)/AppData/abbreviation
+- macOS/Linux: ~/.config/abbreviation
+
+## Added
+
+- `add` command: Adds custom abbreviations to the database.
+- `remove` command: Removes custom abbreviations from the database.
+- `list-user` command: Lists all custom abbreviations from the database.
+- New **LICENSE** file to the repository.
+- Added `sjson` dependency for editing JSON data.
+
+## Changed
+
+- Version command now lists details about the OS and architecture it's running on.
+- Production binaries in GitHub releases are now much more optimized.
+- Help command now includes the new commands and has been *slightly* reformatted.
+- Search command now also searches for user made abbreviations.
+- Modified error messages.
+
 # v0.5.1
 
 ## Bug Fixes